Rename to StandardComboCounter, add HealthDisplay abstraction.
