+++ osu.Game/osu.Game.csproj
@@ -97,6 +97,7 @@
     <Compile Include="Modes\Score.cs" />
     <Compile Include="Modes\ScoreProcesssor.cs" />
     <Compile Include="Modes\UI\HealthDisplay.cs" />
+    <Compile Include="Modes\UI\StandardHealthDisplay.cs" />
     <Compile Include="Online\API\IOnlineComponent.cs" />
     <Compile Include="Online\API\Requests\GetUserRequest.cs" />
     <Compile Include="Overlays\DragBar.cs" />
@@ -167,7 +168,7 @@
     <Compile Include="Screens\Play\PlayerInputManager.cs" />
     <Compile Include="Screens\Play\PlayerLoader.cs" />
     <Compile Include="Screens\Play\SkipButton.cs" />
-    <Compile Include="Modes\UI\ComboCounter.cs" />
+    <Compile Include="Modes\UI\StandardComboCounter.cs" />
     <Compile Include="Modes\UI\StandardHUDOverlay.cs" />
     <Compile Include="Screens\Select\CarouselContainer.cs" />
     <Compile Include="Screens\Select\MatchSongSelect.cs" />
@@ -185,7 +186,7 @@
     <Compile Include="Modes\UI\Playfield.cs" />
     <Compile Include="Modes\UI\HUDOverlay.cs" />
     <Compile Include="Screens\Select\EditSongSelect.cs" />
-    <Compile Include="Modes\UI\BaseComboCounter.cs" />
+    <Compile Include="Modes\UI\ComboCounter.cs" />
     <Compile Include="Modes\UI\ComboResultCounter.cs" />
     <Compile Include="Graphics\UserInterface\RollingCounter.cs" />
     <Compile Include="Graphics\UserInterface\Volume\VolumeControlReceptor.cs" />
